requirements.txt
tkinter
opencv-python
numpy
Pillow
pyzmq
qrcode
pyzbar
pyarmor
pyinstaller
cryptography
